labnogoal: move parser.go into separate mparser package for potential .js replacement
