+++ goal/transpile/mparser/parser.go
@@ -7,7 +7,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package transpile
+package mparser
 
 import (
 	"fmt"
